+++ app/src/main/java/com/planarform/daniel/owstats/OWAverageStats.java
@@ -8,11 +8,11 @@ import java.util.Map;
 
 public class OWAverageStats {
     Double objectiveKillsAverage;
-    Integer damageDoneAverage;
+    Double damageDoneAverage;
     Double meleeFinalBlowsAverage;
     Double deathsAverage;
     Double objectiveTimeAverage;
-    Integer healingDoneAverage;
+    Double healingDoneAverage;
     Double soloKillsAverage;
     Double timeSpentOnFireAverage;
     Double eliminationsAverage;
@@ -20,11 +20,11 @@ public class OWAverageStats {
 
     public OWAverageStats(Map stats) {
         this.objectiveKillsAverage = (Double)stats.get("objective_kills_avg");
-        this.damageDoneAverage = (Integer)stats.get("damage_done_avg");
+        this.damageDoneAverage = (Double)stats.get("damage_done_avg");
         this.meleeFinalBlowsAverage = (Double)stats.get("melee_final_blows_avg");
         this.deathsAverage = (Double)stats.get("deaths_avg");
         this.objectiveTimeAverage = (Double)stats.get("objective_time_avg");
-        this.healingDoneAverage = (Integer)stats.get("healing_done_avg");
+        this.healingDoneAverage = (Double)stats.get("healing_done_avg");
         this.soloKillsAverage = (Double)stats.get("solo_kills_avg");
         this.timeSpentOnFireAverage = (Double)stats.get("time_spent_on_fire_avg");
         this.eliminationsAverage = (Double)stats.get("eliminations_avg");
