+++ app/src/main/java/com/planarform/daniel/owstats/OWMccree.java
@@ -7,17 +7,17 @@ import java.util.Map;
  */
 
 public class OWMccree extends OWHero {
-    Integer deadeyeKillsMostInGame;
-    Integer fanTheHammerKills;
-    Integer deadeye_kills;
-    Integer meleeFinalBlowsMostInGame;
+    Double deadeyeKillsMostInGame;
+    Double fanTheHammerKills;
+    Double deadeye_kills;
+    Double meleeFinalBlowsMostInGame;
 
     public OWMccree (Map stats) {
         super(stats);
-        this.deadeyeKillsMostInGame = (Integer)stats.get("deadeye_kills_most_in_game");
-        this.fanTheHammerKills = (Integer)stats.get("fan_the_hammer_kills");
-        this.deadeye_kills = (Integer)stats.get("deadeye_kills");
-        this.meleeFinalBlowsMostInGame = (Integer)stats.get("melee_final_blows_most_in_game");
+        this.deadeyeKillsMostInGame = (Double)stats.get("deadeye_kills_most_in_game");
+        this.fanTheHammerKills = (Double)stats.get("fan_the_hammer_kills");
+        this.deadeye_kills = (Double)stats.get("deadeye_kills");
+        this.meleeFinalBlowsMostInGame = (Double)stats.get("melee_final_blows_most_in_game");
     }
 
 }
