+++ app/src/main/java/com/planarform/daniel/owstats/OWTracer.java
@@ -7,18 +7,18 @@ import java.util.Map;
  */
 
 public class OWTracer extends OWHero {
-    Integer meleeFinalBlowsMostInGame;
-    Integer pulseBombKills;
-    Integer pulseBombKillsMostInGame;
-    Integer pulseBombsAttached;
-    Integer pulseBombsAttachedMostInGame;
+    Double meleeFinalBlowsMostInGame;
+    Double pulseBombKills;
+    Double pulseBombKillsMostInGame;
+    Double pulseBombsAttached;
+    Double pulseBombsAttachedMostInGame;
 
     public OWTracer(Map stats) {
         super(stats);
-        this.meleeFinalBlowsMostInGame = (Integer)stats.get("melee_final_blows_most_in_game");
-        this.pulseBombKills = (Integer)stats.get("pulse_bomb_kills");
-        this.pulseBombKillsMostInGame = (Integer)stats.get("pulse_bomb_kills_most_in_game");
-        this.pulseBombsAttached = (Integer)stats.get("pulse_bombs_attached");
-        this.pulseBombsAttachedMostInGame = (Integer)stats.get("pulse_bombs_attached_most_in_game");
+        this.meleeFinalBlowsMostInGame = (Double)stats.get("melee_final_blows_most_in_game");
+        this.pulseBombKills = (Double)stats.get("pulse_bomb_kills");
+        this.pulseBombKillsMostInGame = (Double)stats.get("pulse_bomb_kills_most_in_game");
+        this.pulseBombsAttached = (Double)stats.get("pulse_bombs_attached");
+        this.pulseBombsAttachedMostInGame = (Double)stats.get("pulse_bombs_attached_most_in_game");
     }
 }
